backend/requirements.txt
# Web framework
fastapi==0.109.2
uvicorn[standard]==0.27.1

# Database
sqlalchemy==2.0.27
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Settings
pydantic-settings==2.2.1
pydantic==2.6.1

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==42.0.2
PyJWT==2.8.0

# Task Queue
celery==5.3.6
redis==5.0.1
flower==2.0.1  # Celery monitoring UI

# HTTP client
httpx==0.26.0

# Browser automation
playwright==1.41.2

# Platform SDKs
tweepy==4.14.0
spotipy==2.23.0

# File handling
python-multipart==0.0.9
Pillow==10.2.0

# Utilities
python-dateutil==2.8.2
pytz==2024.1

# Monitoring
sentry-sdk[fastapi]==1.40.6
